auto_arima_example/src/requirements.txt
pandas==0.25.0
pyramid-arima==0.8.1
statsmodels==0.10.1